requirements.txt
amqp==2.6.1
APScheduler==3.6.3
asgiref==3.4.1
backports.zoneinfo==0.2.1
billiard==3.6.4.0
cachetools==4.2.2
celery==4.4.7
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
dj-database-url==0.5.0
Django==3.2.6
django-celery-beat==2.2.0
django-cors-headers==3.8.0
django-timezone-field==4.2.1
gunicorn==20.1.0
h11==0.12.0
httptools==0.3.0
idna==3.2
kombu==4.6.11
prompt-toolkit==3.0.19
psycopg2-binary==2.9.1
python-crontab==2.5.1
python-dateutil==2.8.2
python-dotenv==0.19.0
python-telegram-bot==13.7
pytz==2021.1
redis==3.5.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
sseclient-py==1.7
tornado==6.1
tzlocal==3.0
urllib3==1.26.6
uvicorn==0.15.0
uvloop==0.16.0
vine==1.3.0
wcwidth==0.2.5
whitenoise==5.3.0
